ref(ts) Move ActivityItem components to typescript/function components.
I had to adjust a few other components to get the typings to match up.
