+++ src/sentry/static/sentry/app/components/events/userFeedback.tsx
@@ -36,12 +36,15 @@ class EventUserFeedback extends React.Component<Props> {
 
   render() {
     const {className, report} = this.props;
+    const user = report.user
+      ? report.user
+      : {name: report.name, email: report.email, id: '', username: '', ip_address: ''};
 
     return (
       <div className={className}>
         <ActivityItem
           date={report.dateCreated}
-          author={{type: 'user', user: report}}
+          author={{type: 'user', user}}
           header={
             <div>
               <ActivityAuthor>{report.name}</ActivityAuthor>
