+++ src/sentry/static/sentry/app/utils/isRenderFunc.tsx
@@ -1,5 +1,5 @@
 /**
- * Generic type guard for children a function patterns.
+ * Generic type guard for children as a function patterns.
  */
 export function isRenderFunc<T>(func: React.ReactNode | Function): func is T {
   return typeof func === 'function';
